backend/requirements.txt
fastapi
uvicorn
python-dotenv
beautifulsoup4
requests
google-generativeai